requirements.txt
future==0.16.0
python-whois==0.6.4
